sample: Refactor key management
Moves key management to a dedicated key_mgmt file pair, and removes
a number of magic numbers for accessing keys and common values like
tje public key buffer size. Keys now have an enum for the index value
and should be reference via the enum value.

Some refactoring was done for clarity in other files that deal with
key management.

Signed-off-by: Kevin Townsend <kevin.townsend@linaro.org>