Allow better sharing of test certificates and keys
Instead of making these variables statics that are instantiated in every
place they are included, make the header file contain extern
declarations, and include the data in the compiled .c file.

Since sizeof is not helpful for extern declarations (it would be the
size of the pointer), create a helper const that holds the size.

This increases the image by `ptr_size_t` bytes for each certificate or
key.  However, if there are multiple uses, they will be shared instead
of each instance getting a full copy of the key or certificate.

Signed-off-by: David Brown <david.brown@linaro.org>