Implementing provisioning data storate
The provisioning data is stored as individual items in persistent
storage.  Add support for storing this, retrieving it, and waiting for
it to become available on system startup.  Currently, persistent storage
does not persist across reboots on available targets, so the storage is
not yet checked on startup.

Adds code to the caserver query to place the provisioning data into
persistent storage, and the beginnings of code in azure to retrieve the
data.  The azure code does not yet make use of this data, but its
startup is coordinated with it being available.

Signed-off-by: David Brown <david.brown@linaro.org>