doc: Add note about robust protocols
Add a note to the encrypted payload indicating that some protocols don't need
the staggered protocol, if they can guarantee that the CEK messages will always
be delivered.  It is unclear if MQTT has any ordering guarantees, so it will
still be necessary to have the sequence numbers in the messages.

Signed-off-by: David Brown <david.brown@linaro.org>