tfm_secure_partition: Space out the numbering a bit
Put a gap between the service numbers used in our partitions so that
they won't all have to keep being renumbered when one partition gains a
new service.

Note that the vendor ID is still set to 0xFFFFF, which will likely have
to be fixed before these can be combined with any other partitions in an
application.

Signed-off-by: David Brown <david.brown@linaro.org>