sp: Add persistent service IDs to each partition
The partition ID (pid) is used by the key derivation to derive unique
keys.  However, if the pid is not specified in the manifest, the build
system will generate some for it.  This, however, means these IDs can
change across builds.  Assign each partition a fixed ID.

Since we're breaking new ground here, we have decided our vendor ID will
be the ascii for "Li" (the start of Linaro), 0x4c69, the next byte will
be the partition we are describing. We'll leave the lower bits available
so that SIDs and PIDs can match up.

Signed-off-by: David Brown <david.brown@linaro.org>