sp: Match SIDs with the partition IDs
Again, we're breaking new ground here, but we'll use the high 24 bits
of the partition ID as the beginning of the calls for the service ID.
This doesn't leave as much space as the 4k of calls allowed by the
suggested number, but makes it a little easier to map these numbers.  It
is difficult to see a given partition having more than 256 calls.

Signed-off-by: David Brown <david.brown@linaro.org>