provision: Allow selective provisioning update
In order to split the provisioning information into multiple REST
requests, it is necessary to be able to selectively update provisioning
data.  Implement this by adding an enum bit field indicating which
fields should be updated.

By itself, this change will cause provisioning data to no longer be
considered updated, and a subsequent change to the query operations is
needed to go with this.

Signed-off-by: David Brown <david.brown@linaro.org>