key_mgmt: Hide array abstraction a bit
Instead of a zero-argument function that returns an array of contexts
that the caller needs to know the length of, have the `km_key_context`
function take an index argument, and return a direct pointer to that
entry (or null if the index is invalid).  This hides the abstraction
just a little bit, and would allow this to be implemented differently.

Signed-off-by: David Brown <david.brown@linaro.org>