infer: Fix function prototype
Change function declaration `foo()` to `foo()`.  In C, `foo()` is not a
function prototype but merely a declaration of the existence of the
function.  The actual function can be different, and arbitrary arguments
can be passed.  In order to declare a function in C with zero arguments,
it is necessary to say `foo(void)`.

These can be found by adding something like this to the west build
command:

    west build ... -DEXTRA_CFLAGS=-Wstrict-prototypes

Unfortunately, there are a bunch of these in the Zephyr code, so it is
necessary to dig through the output to look for ones affecting our code.

Signed-off-by: David Brown <david.brown@linaro.org>